protein-synthesis-ml/requirements.txt
pandas>=2.1.0
rdkit-pypi>=2022.9.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
chembl-webresource-client>=0.10.0
xgboost>=2.0.0
matplotlib>=3.7.0
pytdc>=1.1.0
joblib>=1.3.0
streamlit>=1.28.0
umap-learn>=0.5.0
pyyaml>=6.0
torch>=2.0.0
transformers>=4.30.0
